1.1.6

+++ src/lgfx/v1/misc/pixelcopy.hpp
@@ -284,7 +284,10 @@ namespace lgfx
 
         int32_t x = param->src_x;
         int32_t y = param->src_y;
-        if (param->src_x == param->src_xe && param->src_y == param->src_ye && static_cast<uint32_t>(param->src_x) < src_width && static_cast<uint32_t>(param->src_y) < src_height)
+        if (param->src_x == param->src_xe
+         && param->src_y == param->src_ye
+         && static_cast<uint32_t>(param->src_x) < static_cast<uint32_t>(src_width)
+         && static_cast<uint32_t>(param->src_y) < static_cast<uint32_t>(src_height))
         {
           uint32_t i = (x + y * src_bitwidth) * src_bits;
           uint32_t raw = (s[i >> 3] >> (-(int32_t)(i + src_bits) & 7)) & src_mask;
@@ -308,8 +311,8 @@ namespace lgfx
             {
               uint32_t rate = rate_x * rate_y;
               argb[4] += rate;
-              if (static_cast<uint32_t>(y) < src_height
-               && static_cast<uint32_t>(x) < src_width)
+              if (static_cast<uint32_t>(y) < static_cast<uint32_t>(src_height)
+               && static_cast<uint32_t>(x) < static_cast<uint32_t>(src_width))
               {
                 uint32_t k = (i + x) * src_bits;
                 uint32_t raw = (s[k >> 3] >> (-(int32_t)(k + src_bits) & 7)) & src_mask;
@@ -376,7 +379,10 @@ namespace lgfx
         int32_t x = param->src_x;
         int32_t y = param->src_y;
         auto color = &s[x + y * src_width];
-        if (param->src_x == param->src_xe && param->src_y == param->src_ye && static_cast<uint32_t>(param->src_x) < src_width && static_cast<uint32_t>(param->src_y) < src_height)
+        if (param->src_x == param->src_xe
+         && param->src_y == param->src_ye
+        && static_cast<uint32_t>(param->src_x) < static_cast<uint32_t>(src_width)
+        && static_cast<uint32_t>(param->src_y) < static_cast<uint32_t>(src_height))
         {
           if (!(*color == param->transp))
           {
@@ -397,8 +403,8 @@ namespace lgfx
             {
               uint32_t rate = rate_x * rate_y;
               argb[4] += rate;
-              if (static_cast<uint32_t>(y) < src_height
-               && static_cast<uint32_t>(x) < src_width
+              if (static_cast<uint32_t>(y) < static_cast<uint32_t>(src_height)
+               && static_cast<uint32_t>(x) < static_cast<uint32_t>(src_width)
                && !(*color == param->transp))
               {
                 if (std::is_same<TSrc, argb8888_t>::value) { rate *= color->A8(); }
